% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_afd_name}
\alias{check_afd_name}
\title{Check if name is valid on the Australian Faunal Directory}
\usage{
check_afd_name(taxon)
}
\arguments{
\item{taxon}{(Character) A query taxon}
}
\value{
(character) A valid taxon name
}
\description{
Check if name is valid on the Australian Faunal Directory
}
