% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_afd_presence}
\alias{check_afd_presence}
\title{Check if taxonomic name is present in the Australian Faunal Directory}
\usage{
check_afd_presence(taxa, quiet = FALSE)
}
\arguments{
\item{taxa}{(Character) A query taxonomic name, or vector of taxonomic names}

\item{quiet}{(Logical) Option to print warnings to console}
}
\value{
(character) A valid taxon name
}
\description{
Check if taxonomic name is present in the Australian Faunal Directory
}
