% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{check_afd_query}
\alias{check_afd_query}
\title{Check query is not too large for the AFD server}
\usage{
check_afd_query(taxon, check_name = TRUE)
}
\arguments{
\item{taxon}{(Character) A query taxon}

\item{check_name}{(Logical) Whether to check the name is valid first using check_afd_name}
}
\value{
a valid URL query, or FALSE if the query is too large
}
\description{
Check query is not too large for the AFD server
}
