% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_afd_checklist}
\alias{fetch_afd_checklist}
\title{Fetch data from the Australian Faunal Directory}
\usage{
fetch_afd_checklist(taxon, retry_attempt = 3, retry_wait = 5, quiet = FALSE)
}
\arguments{
\item{taxon}{(Character) A query taxon}

\item{retry_attempt}{(Integer) The number of query attempts in case of query failure due to poor internet connection.}

\item{retry_wait}{(Integer) How long to wait between query attempts.}

\item{quiet}{(Logical) Whether the function should print progress to console.}
}
\value{
A dataframe containing
}
\description{
Fetch data from the Australian Faunal Directory
}
