% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_afd_csv}
\alias{fetch_afd_csv}
\title{Internal function to download CSV from AFD}
\usage{
fetch_afd_csv(
  taxon,
  retry_attempt = 3,
  retry_wait = 5,
  quiet = FALSE,
  check_name = TRUE
)
}
\arguments{
\item{taxon}{(Character) A query taxon}

\item{retry_attempt}{(Integer) The number of query attempts in case of query failure due to poor internet connection.}

\item{retry_wait}{(Integer) How long to wait between query attempts.}

\item{quiet}{(Logical) Whether the function should print progress to console.}

\item{check_name}{(Logical) Whether to check the name is valid first using check_afd_name}
}
\value{

}
\description{
Internal function to download CSV from AFD
}
