% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_afd_children}
\alias{get_afd_children}
\title{Retrieve the immediate children of a query taxa}
\usage{
get_afd_children(taxon)
}
\arguments{
\item{taxon}{(Character) A query taxon}
}
\value{
a dataframe containing child data
}
\description{
Retrieve the immediate children of a query taxa
}
