% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_afd_rank}
\alias{get_afd_rank}
\title{Get the taxonomic rank of a query on AFD}
\usage{
get_afd_rank(taxon)
}
\arguments{
\item{taxon}{(Character) A query taxon}
}
\value{

}
\description{
Get the taxonomic rank of a query on AFD
}
