% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_afd_children}
\alias{get_afd_children}
\title{Retrieve the immediate children of a query taxa}
\usage{
get_afd_children(taxa, quiet = FALSE)
}
\arguments{
\item{taxa}{(Character) A query taxonomic name, or vector of taxonomic names}

\item{quiet}{(Logical) Option to print warnings to console}
}
\value{
A dataframe containing taxonomic children data for a taxon, or a list of data frames if multiple taxon names were provided.
}
\description{
Retrieve the immediate children of a query taxa
}
