% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_afd_occurance}
\alias{get_afd_occurance}
\title{Get occurance data from the Australian Faunal Directory}
\usage{
get_afd_occurance(taxon, return = "states")
}
\arguments{
\item{taxon}{(Character) A query taxon}

\item{return}{The type of occurance data to return. Options are "states" for just the state occurance data,
or "ibra" for Australia's Bioregions occurance data (see https://www.environment.gov.au/land/nrs/science/ibra)}
}
\value{

}
\description{
Get occurance data from the Australian Faunal Directory
}
